refactor : 판매자 확인 로직 리펙터링 및 테스트용 제거
