+++ src/main/java/com/sparta/itsmine/domain/comment/controller/CommentController.java
@@ -4,9 +4,9 @@ import com.sparta.itsmine.domain.comment.dto.AddCommentResponseDto;
 import com.sparta.itsmine.domain.comment.dto.CommentRequestDto;
 import com.sparta.itsmine.domain.comment.dto.CommentResponseDto;
 import com.sparta.itsmine.domain.comment.service.CommentService;
-import com.sparta.itsmine.domain.security.UserDetailsImpl;
 import com.sparta.itsmine.global.common.HttpResponseDto;
 import com.sparta.itsmine.global.common.ResponseCodeEnum;
+import com.sparta.itsmine.global.security.UserDetailsImpl;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
