--- src/main/java/com/sparta/itsmine/global/exception/comment/QnaExceptionJH.java
@@ -1,11 +1,0 @@
-package com.sparta.itsmine.global.exception.comment;
-
-import com.sparta.itsmine.global.common.ResponseExceptionEnum;
-import lombok.Getter;
-
-@Getter
-public class QnaExceptionJH extends CommentException{
-    public QnaExceptionJH(ResponseExceptionEnum responseCodeEnum) {
-        super(responseCodeEnum);
-    }
-}
