--- src/main/java/com/sparta/itsmine/global/exception/comment/QnaNotFoundExceptionJH.java
@@ -1,9 +1,0 @@
-package com.sparta.itsmine.global.exception.comment;
-
-import com.sparta.itsmine.global.common.ResponseExceptionEnum;
-
-public class QnaNotFoundExceptionJH extends CommentException {
-    public QnaNotFoundExceptionJH(ResponseExceptionEnum responseCodeEnum) {
-        super(responseCodeEnum);
-    }
-}
